requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
faiss-cpu==1.11.0
idna==3.10
numpy==2.2.5
packaging==25.0
PyPDF2==3.0.1
requests==2.32.3
tqdm==4.67.1
urllib3==2.4.0
